pyproject.toml
[tool.poetry]
name = "recipe-summarizer"
version = "0.1.0"
description = ""
authors = ["Thomas Shih <55861476+ThomasShih@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.322"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
cohere = "^4.32"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
html2text = "^2020.1.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
